+++ src/Webform.unit.js
@@ -61,10 +61,36 @@ import columnWithConditionalComponents from '../test/forms/columnWithConditional
 import formWithSurvey from '../test/forms/formWithSurvey';
 import formWithSelectBoxes from '../test/forms/formWithSelectBoxes';
 import formWithDayComp from '../test/forms/formWithDayComp';
+import formWithCalcValue from '../test/forms/formWithCalcValue';
+import testClearOnHideInsideEditGrid from '../test/forms/clearOnHideInsideEditGrid';
 
 /* eslint-disable max-statements */
 describe('Webform tests', function() {
   this.retries(3);
+
+  it('Should recalculate value when submission is being set in edit mode', function(done) {
+    const formElement = document.createElement('div');
+    const form = new Webform(formElement);
+
+    form.setForm(formWithCalcValue).then(() => {
+      const numberComp = form.getComponent('number');
+      const checkbox = form.getComponent('checkbox');
+
+      form.setSubmission({}).then(() => {
+        setTimeout(() => {
+          assert.equal(numberComp.dataValue, 0);
+          assert.equal(checkbox.dataValue, true);
+          form.setSubmission({ data: { number: 7, checkbox: true } }).then(() => {
+            setTimeout(() => {
+              assert.equal(numberComp.dataValue, 7);
+              assert.equal(checkbox.dataValue, false);
+              done();
+            }, 500);
+          });
+        }, 500);
+      });
+    }).catch((err) => done(err));
+  });
 
   it('Should show survey values in html render mode', function(done) {
     const formElement = document.createElement('div');
@@ -2326,12 +2352,17 @@ describe('Webform tests', function() {
               dataSourceDisplay: 'some value'
             },
           state: 'submitted'
-        });
-        setTimeout(() => {
-          const dataSourceDisplay = form.getComponent('dataSourceDisplay');
-          assert.equal(dataSourceDisplay.dataValue, 'some value', 'Should set and keep the value');
-          done();
-        }, 1000);
+        })
+        .then(() => {
+          const dataSource = form.getComponent('datasource');
+          dataSource.dataValue = { value: 'some value' };
+          form.checkData(null, { dataSourceInitialLoading: true });
+          setTimeout(() => {
+            const dataSourceDisplay = form.getComponent('dataSourceDisplay');
+            assert.equal(dataSourceDisplay.dataValue, 'some value', 'Should set and keep the value');
+            done();
+          }, 1000);
+       });
       }).catch(done);
     });
     it('Should calculate value properly in editing mode', (done) => {
@@ -3014,6 +3045,55 @@ describe('Webform tests', function() {
             assert.equal(radioConditional.visible, true, 'Should stay visible');
 
             done();
+          }, 250);
+        }, 250);
+      }).catch(done);
+    });
+
+    it('Should execute clearOnHide if visibility of the component inside an EditGrid has changed', (done) => {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement, { language: 'en', template: 'bootstrap3' });
+
+      form.setForm(testClearOnHideInsideEditGrid).then(() => {
+        form.submission = {
+          state: 'submitted',
+          data: {
+            subsidiaryEditGrid: [
+              {
+                subsidiaryEntityContainer: {
+                  entityFullName: 'test',
+                  divisionNum: '',
+                  entityType: 'otherEntity',
+                  ifOtherEntityPleaseExplain: 'test',
+                },
+              },
+            ],
+          },
+        };
+
+        setTimeout(() => {
+          const clearOnHideField = form.getComponent([
+            'subsidiaryEditGrid',
+            0,
+            'subsidiaryEntityContainer',
+            'ifOtherEntityPleaseExplain',
+          ]);
+          const radioTrigger = form.getComponent(['subsidiaryEditGrid', 0, 'subsidiaryEntityContainer', 'entityType']);
+          assert.equal(form.rootPristine, true, 'Should not change this prop  after setting a submission');
+          assert.equal(clearOnHideField.visible, true, 'Should become visible');
+          assert.equal(clearOnHideField.dataValue, 'test', 'Should set a value from  the submission');
+
+          radioTrigger.setValue('subsidiary', { modified: true });
+          setTimeout(() => {
+            assert.equal(clearOnHideField.visible, false, 'Should become invisible');
+
+            radioTrigger.setValue('otherEntity', { modified: true });
+            setTimeout(() => {
+              assert.equal(clearOnHideField.visible, true, 'Should become visible');
+              assert.equal(clearOnHideField.dataValue, '', 'Should clear a value due to the clearOnHide');
+
+              done();
+            }, 250);
           }, 250);
         }, 250);
       }).catch(done);