+++ src/components/_classes/component/Component.js
@@ -4,15 +4,11 @@ import NativePromise from 'native-promise-only';
 import Tooltip from 'tooltip.js';
 import _ from 'lodash';
 import isMobile from 'ismobilejs';
-import Formio from '../../../Formio';
+import { GlobalFormio as Formio } from '../../../Formio';
 import * as FormioUtils from '../../../utils/utils';
 import Validator from '../../../validator/Validator';
-import Templates from '../../../templates/Templates';
 import {
-  fastCloneDeep,
-  boolValue,
-  getDataParentComponent,
-  getComponentPath
+  fastCloneDeep, boolValue, getComponentPath, isInsideScopingComponent,
 } from '../../../utils/utils';
 import Element from '../../../Element';
 import ComponentModal from '../componentModal/ComponentModal';
@@ -30,6 +26,12 @@ const QUILL_URL = isIEBrowser
   : 'https://cdn.quilljs.com/2.0.0-dev.3';
 const QUILL_TABLE_URL = 'https://cdn.form.io/quill/quill-table.js';
 const ACE_URL = 'https://cdn.form.io/ace/1.4.10/ace.js';
+
+let Templates = Formio.Templates;
+
+if (!Templates) {
+  Templates = require('../../../templates/Templates').default;
+}
 
 /**
  * This is the Component class
@@ -873,7 +875,7 @@ export default class Component extends Element {
   getFormattedTooltip(tooltipValue) {
     const tooltip = this.interpolate(tooltipValue || '').replace(/(?:\r\n|\r|\n)/g, '<br />');
 
-    return tooltip ? this.t(tooltip, { _userInput: true }) : '';
+    return tooltip ? this.t(tooltip, { _userInput: true }).replace(/"/g, '&quot;') : '';
   }
 
   isHtmlRenderMode() {
@@ -1039,8 +1041,8 @@ export default class Component extends Element {
     }
   }
 
-  setOpenModalElement() {
-    this.componentModal.setOpenModalElement(this.getModalPreviewTemplate());
+  setOpenModalElement(template) {
+    this.componentModal.setOpenModalElement(template || this.getModalPreviewTemplate());
   }
 
   getModalPreviewTemplate() {
@@ -1123,8 +1125,11 @@ export default class Component extends Element {
     if (!this.builderMode && !this.previewMode && this.component.modalEdit) {
       const modalShouldBeOpened = this.componentModal ? this.componentModal.isOpened : false;
       const currentValue = modalShouldBeOpened ? this.componentModal.currentValue : this.dataValue;
+      const openModalTemplate = this.componentModal && modalShouldBeOpened
+        ? this.componentModal.openModalTemplate
+        : null;
       this.componentModal = this.createComponentModal(element, modalShouldBeOpened, currentValue);
-      this.setOpenModalElement();
+      this.setOpenModalElement(openModalTemplate);
     }
 
     this.attached = true;
@@ -2022,7 +2027,7 @@ export default class Component extends Element {
     // clearOnHide defaults to true for old forms (without the value set) so only trigger if the value is false.
     if (
       // if change happens inside EditGrid's row, it doesn't trigger change on the root level, so rootPristine will be true
-      (!this.rootPristine || this.options.server || getDataParentComponent(this)?.hasScopedChildren) &&
+      (!this.rootPristine || this.options.server || isInsideScopingComponent(this)) &&
       this.component.clearOnHide !== false &&
       !this.options.readOnly &&
       !this.options.showHiddenFields
@@ -2381,7 +2386,7 @@ export default class Component extends Element {
   }
 
   get shouldAddDefaultValue() {
-    return !this.options.noDefaults || this.component.defaultValue || this.component.customDefaultValue;
+    return !this.options.noDefaults || (this.component.defaultValue && !this.isEmpty(this.component.defaultValue)) || this.component.customDefaultValue;
   }
 
   get defaultValue() {
@@ -2471,12 +2476,13 @@ export default class Component extends Element {
       return changed;
     }
     const isArray = Array.isArray(value);
+    const valueInput = this.refs.fileLink || this.refs.input;
     if (
       isArray &&
       Array.isArray(this.defaultValue) &&
       this.refs.hasOwnProperty('input') &&
-      this.refs.input &&
-      (this.refs.input.length !== value.length) &&
+      valueInput &&
+      (valueInput.length !== value.length) &&
       this.visible
     ) {
       this.redraw();
@@ -2678,6 +2684,7 @@ export default class Component extends Element {
     // hidden and set to clearOnHide (Don't calculate a value for a hidden field set to clear when hidden)
     const { hidden, clearOnHide } = this.component;
     const shouldBeCleared = (!this.visible || hidden) && clearOnHide && !this.rootPristine;
+    const allowOverride = _.get(this.component, 'allowCalculateOverride', false);
 
     // Handle all cases when calculated values should not fire.
     if (
@@ -2685,7 +2692,7 @@ export default class Component extends Element {
       !(this.component.calculateValue || this.component.calculateValueVariable) ||
       shouldBeCleared ||
       (this.options.server && !this.component.calculateServer) ||
-      flags.dataSourceInitialLoading
+      (flags.dataSourceInitialLoading && allowOverride)
     ) {
       return false;
     }
@@ -2700,14 +2707,8 @@ export default class Component extends Element {
 
     const changed = !_.isEqual(dataValue, calculatedValue);
 
-    if (flags.fromSubmission && this.component.persistent === true) {
-      // If we set value from submission and it differs from calculated one, set the calculated value to prevent overriding dataValue in the next pass
-      this.calculatedValue = calculatedValue;
-      return false;
-    }
-
     // Do not override calculations on server if they have calculateServer set.
-    if (this.component.allowCalculateOverride) {
+    if (allowOverride) {
       const firstPass = (this.calculatedValue === undefined);
       if (firstPass) {
         this.calculatedValue = null;
@@ -2724,6 +2725,12 @@ export default class Component extends Element {
       if (flags.isReordered || !calculationChanged) {
         return false;
       }
+
+    if (flags.fromSubmission && this.component.persistent === true) {
+      // If we set value from submission and it differs from calculated one, set the calculated value to prevent overriding dataValue in the next pass
+      this.calculatedValue = calculatedValue;
+      return false;
+    }
 
       // If this is the firstPass, and the dataValue is different than to the calculatedValue.
       if (firstPass && !this.isEmpty(dataValue) && changed && calculationChanged) {
@@ -3002,6 +3009,18 @@ export default class Component extends Element {
 
     const hasErrors = !!messages.filter(message => message.level === 'error').length;
 
+    let invalidInputRefs = inputRefs;
+    if (this.component.multiple) {
+      const inputRefsArray = Array.from(inputRefs);
+      inputRefsArray.forEach((input) => {
+        this.setElementInvalid(this.performInputMapping(input), false);
+      });
+      invalidInputRefs = inputRefsArray.filter((ref) => {
+        return messages.some?.((msg) => {
+          return msg.context.input === ref;
+        });
+      });
+    }
     if (messages.length) {
       if (this.refs.messageContainer) {
         this.empty(this.refs.messageContainer);
@@ -3013,9 +3032,9 @@ export default class Component extends Element {
         external: !!external,
       };
       this.emit('componentError', this.error);
-      this.addMessages(messages, dirty, inputRefs);
-      if (inputRefs) {
-        this.setErrorClasses(inputRefs, dirty, hasErrors, !!messages.length);
+      this.addMessages(messages, dirty, invalidInputRefs);
+      if (invalidInputRefs) {
+        this.setErrorClasses(invalidInputRefs, dirty, hasErrors, !!messages.length);
       }
     }
     else if (this.error && this.error.external === !!external) {
@@ -3026,8 +3045,8 @@ export default class Component extends Element {
         this.empty(this.refs.modalMessageContainer);
       }
       this.error = null;
-      if (inputRefs) {
-        this.setErrorClasses(inputRefs, dirty, hasErrors, !!messages.length);
+      if (invalidInputRefs) {
+        this.setErrorClasses(invalidInputRefs, dirty, hasErrors, !!messages.length);
       }
       this.clearErrorClasses();
     }