+++ src/components/select/fixtures/comp15.js
@@ -1,31 +1,39 @@
 export default {
-	_id: '60ec032f86daf81238ab4e37',
-	type: 'form',
-	components: [
-		{
-			label: 'Select',
-			widget: 'choicesjs',
-			tableView: true,
-			dataSrc: 'custom',
-      valueProperty:'value',
-			data: {
-				custom:"values = Promise.resolve([\n  {label: 'A', value: 'aa'}, \n  {label: 'B', value: 'bb'}, \n  {label: 'C', value: 'cc'}\n]);",
-			},
-			key: 'select',
-			type: 'select',
-			input: true
-		},
-		{
-			type: 'button',
-			label: 'Submit',
-			key: 'submit',
-			disableOnInvalid: true,
-			input: true,
-			tableView: false
-		}
-	],
-	title: 'custom async',
-	display: 'form',
-	name: 'customAsync',
-	path: 'customasync',
+  _id: '6110fab8ca845631779c2bc5',
+  type: 'form',
+  components: [
+    {
+      label: 'Select',
+      widget: 'html5',
+      tableView: true,
+      dataSrc: 'resource',
+      data: {
+        resource: '6110fa7961e298191332514d',
+        project: 'gjrwbdsqkpxesha',
+      },
+      dataType: 'object',
+      valueProperty: 'data',
+      template: '<span>{{ item.data }}</span>',
+      validate: { select: false },
+      key: 'select',
+      type: 'select',
+      searchField: 'data__regex',
+      input: true,
+      addResource: false,
+      reference: false,
+    },
+    {
+      label: 'Submit',
+      showValidations: false,
+      tableView: false,
+      key: 'submit',
+      type: 'button',
+      input: true,
+    },
+  ],
+  title: 'FIO-3011',
+  display: 'form',
+  name: 'fio3011',
+  path: 'fio3011',
+  project: '6108d710d447e01ac3d81b8f',
 };