+++ src/components/tree/Tree.js
@@ -432,7 +432,7 @@ export default class TreeComponent extends NestedDataComponent {
   }
 
   setRoot() {
-    const value = this.dataValue;
+    const value = this.getValue();
     this.treeRoot = new Node(null, value, {
       isNew: this.builderMode ? true : !value.data,
       createComponents: this.createComponents.bind(this),
@@ -448,7 +448,7 @@ export default class TreeComponent extends NestedDataComponent {
   }
 
   getValue() {
-    return this.dataValue;
+    return this.dataValue || {};
   }
 
   updateTree() {