+++ src/utils/utils.js
@@ -1451,7 +1451,6 @@ export function getDataParentComponent(componentInstance) {
   }
 }
 
-
 /**
  * Returns whether the value is a promise
  * @param value
@@ -1464,13 +1463,33 @@ export function getDataParentComponent(componentInstance) {
      && value?.constructor?.name === 'Promise';
  }
 
+/**
+ * Determines if the component has a scoping parent in tree (a component which scopes its children and manages its
+ * changes by itself, e.g. EditGrid)
+ * @param componentInstance
+ * @param firstPass
+ * @returns {boolean|boolean|*}
+ */
+export function isInsideScopingComponent(componentInstance, firstPass = true) {
+  if (!firstPass && componentInstance?.hasScopedChildren) {
+    return true;
+  }
+  const dataParent = getDataParentComponent(componentInstance);
+  if (dataParent?.hasScopedChildren) {
+    return true;
+  }
+  else if (dataParent?.parent) {
+    return isInsideScopingComponent(dataParent.parent, false);
+  }
+  return false;
+}
+
 export function getFocusableElements(element) {
   const focusableSelector =
     `button:not([disabled]), input:not([disabled]), select:not([disabled]),
     textarea:not([disabled]), button:not([disabled]), [href]`;
   return element.querySelectorAll(focusableSelector);
 }
-
 
 // Export lodash to save space with other libraries.
 export { _ };